requirements.txt
Django==4.1.7
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-debug-toolbar==4.0.0